backend/requirements.txt
Flask
markdownify
transformers
pix2tex
beautifulsoup4
ebooklib
Pillow
genanki
ollama